+++ mod/src/RPPostGameStarted.lua
@@ -33,7 +33,6 @@ function RPPostGameStarted:Main(saveState)
     roomIndex = level:GetCurrentRoomIndex()
   end
   local seeds = game:GetSeeds()
-  local player = game:GetPlayer(0)
   local isaacFrameCount = Isaac.GetFrameCount()
 
   Isaac.DebugString("MC_POST_GAME_STARTED")
@@ -170,8 +169,8 @@ function RPPostGameStarted:Main(saveState)
   end
 
   -- Make sure that the festive hat shows
-  -- (enabled until Thanksgiving ends)
-  player:AddNullCostume(16) -- Corresponds to "n016_Christmas.anm2" in the "costumes2.xml" file
+  -- (disabled until Christmas)
+  --player:AddNullCostume(16) -- Corresponds to "n016_Christmas.anm2" in the "costumes2.xml" file
 
   -- Call PostNewLevel manually (they get naturally called out of order)
   RPPostNewLevel:NewLevel()