+++ mod/src/RPPostNewRoom.lua
@@ -4,12 +4,13 @@ local RPPostNewRoom = {}
 -- Includes
 --
 
-local RPGlobals    = require("src/rpglobals")
-local RPFastClear  = require("src/rpfastclear")
-local RPFastTravel = require("src/rpfasttravel")
-local RPSpeedrun   = require("src/rpspeedrun")
-local RPSprites    = require("src/rpsprites")
-local SamaelMod    = require("src/rpsamael")
+local RPGlobals     = require("src/rpglobals")
+local RPFastClear   = require("src/rpfastclear")
+local RPFastTravel  = require("src/rpfasttravel")
+local RPSpeedrun    = require("src/rpspeedrun")
+local RPSprites     = require("src/rpsprites")
+local RPSeededDeath = require("src/rpseededdeath")
+local SamaelMod     = require("src/rpsamael")
 
 -- ModCallbacks.MC_POST_NEW_ROOM (19)
 function RPPostNewRoom:Main()
@@ -522,6 +523,9 @@ function RPPostNewRoom:Race()
 
   -- Remove the final place graphic if it is showing
   RPSprites:Init("place2", 0)
+
+  -- Check for the special death condition
+  RPSeededDeath:PostNewRoom()
 
   -- Prevent players from skipping a floor by using the I AM ERROR room on Womb 2 on the "Everything" race goal
   if stage == LevelStage.STAGE4_2 and -- 8